Started development on level editor.
Level editor is a feature in Colorox that I really wanted to make but before two years didn't knew how to. But now it has been developing. It would not only be a great feature and way for players to create and share their colorox levels but also for me to create levels easily. I am looking forward to it.