Started work on Level Publishing
I read some articles and watched some youtube videos to know and learn about JSON with unity. Level data can be converted to JSON and then stored in a Database. I am integrating Firebase Database as I find it to be simple and easy to operate and understand. Basic UI and Animation and put-together-code are setted up. Only the technical part is left which would be completed easily due to strong base. I'm excited. :)