Put the cherry on top (maybe)
Okay, I started creating the level complete system, if all collectors are filled, a message will pop up in console saying "levelCompleted". I will actual-implement this later. Just wanted to finish this to give the game a finishing feel (atleast for me). Also I encountered a minor bug that looked major at first in level editor, but I fixed it. All commits after this might not be in Level Editor branch but something else to create further content for the game.