Sliders are successfully implemented.
I cannot believe the flexibility of sliders now. In previous colorox, they were buggy because they were just a button on a slider, the game couldn't figure out whether the slider was sliding or button was being pressed which caused lack of control. But now, with manually created buttons, I configured the sliders such that if slightly the slider was moved, it would still shoot energy which is a flexible control. Also, sliders are added to Level Creator and it took like a day.